#251 [fix]-layout테스트
